[REF] l10n_tr,l10n_tr_nilvera_einvoice: move tax office to l10n_tr
Tax Office data is currently owned by l10n_tr_nilvera_einvoice but is
needed across the TR localization (statements, payroll, accounting
reports) without dragging the e-invoice dependency.

Changes:
- Move `l10n_tr_nilvera_einvoice.tax.office` model, its 1046 data
  records, ACLs, and its list/form/action/menu to `l10n_tr`
- Move `l10n_tr_tax_office_id` field on res.partner / res.company
  with its display-name compute, helpers, and view inheritances
- Rename `_get_tax_office_for_edispatch` to `_get_tax_office_name`;
  update the single caller in l10n_tr_nilvera_edispatch
- Drop l10n_tr_nilvera_edispatch's dead `self.ref` fallback overrides
  (the field is now always available via the dep chain)

task-6189470
