+++ addons/l10n_tr/__manifest__.py
@@ -20,11 +20,17 @@ Activates:
     'author': 'Odoo S.A., Drysharks Consulting and Trading Ltd.',
     'depends': [
         'account',
+        'contacts',
     ],
     'auto_install': ['account'],
     'data': [
+        'security/ir.access.csv',
         'data/account_tax_report_data.xml',
         'data/account_tax_report_stamp_tax.xml',
+        'data/l10n_tr.tax.office.csv',
+        'views/l10n_tr_tax_office_views.xml',
+        'views/res_company_views.xml',
+        'views/res_partner_views.xml',
     ],
     'demo': [
         'demo/demo_company.xml',
