+++ addons/l10n_tr/models/__init__.py
@@ -1,6 +1,8 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 from . import account_journal
 from . import account_move_line
+from . import l10n_tr_tax_office
 from . import product
-from . import template_tr
+from . import res_company
 from . import res_partner
+from . import template_tr
