+++ addons/l10n_tr/security/ir.access.csv
@@ -1,0 +1,5 @@
+id,name,model_id,group_id/id,operation,domain
+access_l10n_tr_tax_office_public,l10n_tr_tax_office_group_public_user,l10n_tr.tax.office,base.group_public,r,
+access_l10n_tr_tax_office_portal,l10n_tr_tax_office_group_protal_user,l10n_tr.tax.office,base.group_portal,r,
+access_l10n_tr_tax_office_employee,l10n_tr_tax_office_group_user,l10n_tr.tax.office,base.group_user,r,
+access_l10n_tr_tax_office_group_partner_manager,l10n_tr_tax_office_group_partner_manager,l10n_tr.tax.office,base.group_partner_manager,crud,
