+++ addons/l10n_tr_nilvera_einvoice/demo/demo_company.xml
@@ -2,16 +2,16 @@
 <odoo>
     <record id="base.partner_demo_company_tr" model="res.partner">
         <field name="additional_identifiers" eval="{'TR_MERSIS': '0-123456780100019'}"/>
-        <field name="l10n_tr_tax_office_id" ref="l10n_tr_nilvera_einvoice.tax_office_1045"/>
+        <field name="l10n_tr_tax_office_id" ref="l10n_tr.tax_office_1045"/>
     </record>
 
     <record id="l10n_tr.l10n_tr_res_partner_1" model="res.partner">
         <field name="additional_identifiers" eval="{'TR_MERSIS': '0-123456780200020'}"/>
-        <field name="l10n_tr_tax_office_id" ref="l10n_tr_nilvera_einvoice.tax_office_763"/>
+        <field name="l10n_tr_tax_office_id" ref="l10n_tr.tax_office_763"/>
     </record>
 
     <record id="l10n_tr.l10n_tr_res_partner_2" model="res.partner">
         <field name="additional_identifiers" eval="{'TR_MERSIS': '0-424519348000017'}"/>
-        <field name="l10n_tr_tax_office_id" ref="l10n_tr_nilvera_einvoice.tax_office_545"/>
+        <field name="l10n_tr_tax_office_id" ref="l10n_tr.tax_office_545"/>
     </record>
 </odoo>
