+++ addons/l10n_tr_nilvera_einvoice/models/__init__.py
@@ -5,7 +5,6 @@ from . import account_move_line
 from . import account_move_send
 from . import account_tax
 from . import l10n_tr_nilvera_einvoice_account_tax_code
-from . import l10n_tr_nilvera_einvoice_tax_office
 from . import product_product
 from . import product_template
 from . import res_company
