+++ addons/l10n_tr_nilvera_einvoice/security/ir.access.csv
@@ -1,8 +1,4 @@
 id,name,model_id,group_id/id,operation,domain
 access_l10n_tr_nilvera_einvoice_account_tax_code_readonly,l10n_tr_nilvera_einvoice_account_tax_code_readonly,l10n_tr_nilvera_einvoice.account.tax.code,account.group_account_readonly,r,
 access_l10n_tr_nilvera_einvoice_account_tax_code,l10n_tr_nilvera_einvoice_account_tax_code,l10n_tr_nilvera_einvoice.account.tax.code,account.group_account_basic,crud,
-access_l10n_tr_nilvera_einvoice_tax_office_public,l10n_tr_nilvera_einvoice_tax_office_group_public_user,l10n_tr_nilvera_einvoice.tax.office,base.group_public,r,
-access_l10n_tr_nilvera_einvoice_tax_office_portal,l10n_tr_nilvera_einvoice_tax_office_group_protal_user,l10n_tr_nilvera_einvoice.tax.office,base.group_portal,r,
-access_l10n_tr_nilvera_einvoice_tax_office_employee,l10n_tr_nilvera_einvoice_tax_office_group_user,l10n_tr_nilvera_einvoice.tax.office,base.group_user,r,
-access_l10n_tr_nilvera_einvoice_tax_office_group_partner_manager,l10n_tr_nilvera_einvoice_tax_office_group_partner_manager,l10n_tr_nilvera_einvoice.tax.office,base.group_partner_manager,crud,
 access_l10n_tr_ticarifatura_response_wizard_group_invoicing,l10n_tr_ticarifatura_response_wizard_group_invoicing,l10n_tr.ticarifatura.response.wizard,account.group_account_invoice,crud,
